camhook: split cam-detect code to different file
