+++ src/main/camhook/cam.h
@@ -3,6 +3,13 @@
 
 #include "camhook/config-cam.h"
 
+enum camhook_version {
+    CAMHOOK_OLD,
+    CAMHOOK_NEW,
+};
+
+void camhook_set_version(enum camhook_version version);
+
 void camhook_init(struct camhook_config_cam *config_cam);
 
 void camhook_fini(void);
